+++ Runtime/GetComponentAttribute.cs
@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Diagnostics;
+using SaintsField.Playa;
 using UnityEngine;
 
 namespace SaintsField
 {
     [Conditional("UNITY_EDITOR")]
-    public class GetComponentAttribute: PropertyAttribute, ISaintsAttribute
+    public class GetComponentAttribute: PropertyAttribute, ISaintsAttribute, IPlayaAttribute, IPlayaArraySizeAttribute
     {
         public SaintsAttributeType AttributeType => SaintsAttributeType.Other;
         public string GroupBy { get; }