+++ Samples~/Scripts/Dummy.cs
@@ -2,8 +2,9 @@
 
 namespace SaintsField.Samples.Scripts
 {
-    public class Dummy: MonoBehaviour
+    public class Dummy: MonoBehaviour, IDummy
     {
         public string comment;
+        public string GetComment() => comment;
     }
 }