+++ Samples~/Scripts/IDummy.cs
@@ -1,0 +1,7 @@
+namespace SaintsField.Samples.Scripts
+{
+    public interface IDummy
+    {
+        public string GetComment();
+    }
+}