+++ Samples~/Scripts/SaintsEditor/Issue45/GetComponentArray.cs
@@ -1,0 +1,35 @@
+using System;
+using System.Collections.Generic;
+
+namespace SaintsField.Samples.Scripts.SaintsEditor.Issue45
+{
+    public class GetComponentArray : SaintsMonoBehavior
+    {
+        [GetComponent, PostFieldRichLabel(nameof(DummyNumber), isCallback: true)] public Dummy[] getComponentArray;
+        [GetComponent, PostFieldRichLabel(nameof(DummyNumber), isCallback: true)] public List<Dummy> getComponentList;
+
+        [Serializable]
+        public class GeneralInterface : SaintsInterface<UnityEngine.Object, IDummy> { }
+
+        [GetComponent, PostFieldRichLabel(nameof(DummyNumberI), isCallback: true)]
+        public GeneralInterface[] getComponentIArray;
+
+        [GetComponent, PostFieldRichLabel(nameof(DummyNumberG), isCallback: true)]
+        public List<SaintsInterface<UnityEngine.Object, IDummy>> getComponentIList;
+
+        private string DummyNumber(Dummy dummy)
+        {
+            return dummy? $"{dummy.comment}": "";
+        }
+
+        private string DummyNumberI(GeneralInterface dummyInter)
+        {
+            return dummyInter == null? "":  $"{dummyInter.I.GetComment()}";
+        }
+
+        private string DummyNumberG(SaintsInterface<UnityEngine.Object, IDummy> dummyInter)
+        {
+            return dummyInter == null? "":  $"{dummyInter.I.GetComment()}";
+        }
+    }
+}