+++ Samples~/Scripts/SaintsEditor/Issue45/GetComponentInChildrenArray.cs
@@ -1,0 +1,16 @@
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace SaintsField.Samples.Scripts.SaintsEditor.Issue45
+{
+    public class GetComponentInChildrenArray : SaintsMonoBehavior
+    {
+        [GetComponentInChildren, PostFieldRichLabel(nameof(DummyNumber), isCallback: true)] public Dummy[] getComponentInChildrenArray;
+        [GetComponentInChildren, PostFieldRichLabel(nameof(DummyNumber), isCallback: true)] public List<Dummy> getComponentInChildrenList;
+
+        private string DummyNumber(Dummy dummy)
+        {
+            return dummy? $"{dummy.comment}": "";
+        }
+    }
+}