+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 2.  Fix [#42](https://github.com/TylerTemp/SaintsField/issues/42) GetComponent etc need better error handling and report
 3.  Fix callback won't fill null parameter value when the value of the target is null
 4.  Deprecated `PlayaArraySize`. Just use `ArraySize` instead.
+5.  Fix `SaintsInterface` drawer did not work inside array.
 
 ## 3.1.3